Added option to cleanup or not installation files

+++ manifests/installdb.pp
@@ -27,6 +27,7 @@ define oradb::installdb(
   $puppetDownloadMntPoint  = undef,
   $remoteFile              = true,
   $cluster_nodes           = undef,
+  $cleanup_installfiles    = true,
 )
 {
   if ( $createUser == true ){
@@ -222,35 +223,37 @@ define oradb::installdb(
     }
 
     # cleanup
-    if ( $zipExtract ) {
-      exec { "remove oracle db extract folder ${title}":
-        command => "rm -rf ${downloadDir}/${file}",
-        user    => 'root',
-        group   => 'root',
-        path    => $execPath,
-        cwd     => $oracleBase,
-        require => [Exec["install oracle database ${title}"],
-                    Exec["run root.sh script ${title}"],],
-      }
+    if ( $cleanup_installfiles ) {
+      if ( $zipExtract ) {
+        exec { "remove oracle db extract folder ${title}":
+          command => "rm -rf ${downloadDir}/${file}",
+          user    => 'root',
+          group   => 'root',
+          path    => $execPath,
+          cwd     => $oracleBase,
+          require => [Exec["install oracle database ${title}"],
+                      Exec["run root.sh script ${title}"],],
+          }
 
-      if ( $remoteFile == true ){
-        exec { "remove oracle db file1 ${file1} ${title}":
-          command => "rm -rf ${downloadDir}/${file1}",
-          user    => 'root',
-          group   => 'root',
-          path    => $execPath,
-          cwd     => $oracleBase,
-          require => [Exec["install oracle database ${title}"],
-                      Exec["run root.sh script ${title}"],],
-        }
-        exec { "remove oracle db file2 ${file2} ${title}":
-          command => "rm -rf ${downloadDir}/${file2}",
-          user    => 'root',
-          group   => 'root',
-          path    => $execPath,
-          cwd     => $oracleBase,
-          require => [Exec["install oracle database ${title}"],
-                      Exec["run root.sh script ${title}"],],
+        if ( $remoteFile == true ){
+          exec { "remove oracle db file1 ${file1} ${title}":
+            command => "rm -rf ${downloadDir}/${file1}",
+            user    => 'root',
+            group   => 'root',
+            path    => $execPath,
+            cwd     => $oracleBase,
+            require => [Exec["install oracle database ${title}"],
+                          Exec["run root.sh script ${title}"],],
+          }
+          exec { "remove oracle db file2 ${file2} ${title}":
+            command => "rm -rf ${downloadDir}/${file2}",
+            user    => 'root',
+            group   => 'root',
+            path    => $execPath,
+            cwd     => $oracleBase,
+            require => [Exec["install oracle database ${title}"],
+                        Exec["run root.sh script ${title}"],],
+          }
         }
       }
     }
